refactor partition processing into separate files
