+++ lib/index.ts
@@ -16,3 +16,5 @@ export * from "./box-sides/mergeBoxSideSubgraphs"
 export * from "./box-sides/findIsolatedBoxSides"
 export * from "./renetwork/renetworkWithCondition"
 export * from "./debug/convertFlatBpcToGraphics"
+export * from "./partition-processing/getCanonicalRightFacingGraph"
+export * from "./partition-processing/SchematicPartitionProcessor"
