+++ package.json
@@ -16,9 +16,11 @@
     "@biomejs/biome": "^2.0.0",
     "@react-hook/resize-observer": "^2.0.2",
     "@types/bun": "latest",
+    "@types/debug": "^4.1.12",
     "@vitejs/plugin-react": "^4.5.2",
     "bun-match-svg": "^0.0.11",
     "circuit-to-svg": "^0.0.164",
+    "debug": "^4.4.1",
     "graphics-debug": "^0.0.60",
     "ml-matrix": "^6.12.1",
     "react-cosmos": "^7.0.0",
